🐛extract constants to avoid almost empty recorder chunk
