+++ packages/rum/src/boot/lazyLoadRecorder.spec.ts
@@ -9,7 +9,7 @@ import type { CreateDeflateWorker } from '../domain/deflate'
 import { resetDeflateWorkerState } from '../domain/deflate'
 import { MockWorker } from '../../test'
 import * as replayStats from '../domain/replayStats'
-import { RECORDER_INIT_METRICS_TELEMETRY_NAME } from '../domain/startRecorderInitTelemetry'
+import { RECORDER_INIT_METRICS_TELEMETRY_NAME } from '../domain/replayTelemetryConstants'
 import { makeRecorderApi } from './recorderApi'
 import type { StartRecording } from './postStartStrategy'
 import { lazyLoadRecorder } from './lazyLoadRecorder'
