+++ packages/rum/src/boot/recorderApi.spec.ts
@@ -21,7 +21,8 @@ import type { CreateDeflateWorker } from '../domain/deflate'
 import { resetDeflateWorkerState } from '../domain/deflate'
 import { MockWorker } from '../../test'
 import * as replayStats from '../domain/replayStats'
-import { type RecorderInitMetrics, RECORDER_INIT_METRICS_TELEMETRY_NAME } from '../domain/startRecorderInitTelemetry'
+import { type RecorderInitMetrics } from '../domain/startRecorderInitTelemetry'
+import { RECORDER_INIT_METRICS_TELEMETRY_NAME } from '../domain/replayTelemetryConstants'
 import { makeRecorderApi } from './recorderApi'
 import type { StartRecording } from './postStartStrategy'
