+++ packages/rum/src/domain/replayTelemetryConstants.ts
@@ -1,0 +1,2 @@
+export const RECORDER_INIT_METRICS_TELEMETRY_NAME = 'Recorder init metrics'
+export const SEGMENT_METRICS_TELEMETRY_NAME = 'Segment network request metrics'
