+++ packages/rum/src/domain/segmentCollection/index.ts
@@ -1,4 +1,4 @@
 export { startSegmentCollection, setSegmentBytesLimit } from './segmentCollection'
 export { SEGMENT_BYTES_LIMIT } from './segmentCollection'
 export type { BrowserSegmentMetadataAndSegmentSizes, ReplayPayload } from './buildReplayPayload'
-export { startSegmentTelemetry, SEGMENT_METRICS_TELEMETRY_NAME } from './startSegmentTelemetry'
+export { startSegmentTelemetry } from './startSegmentTelemetry'
